HEXDEV-645: Also shrink the domains of the examplar member frames.
